# ml-service/app/schemas.py
"""Pydantic request/response models for the ml-service API."""

from typing import Literal, Optional

from pydantic import BaseModel, Field


class SensorReadingIn(BaseModel):
    """Mirrors the fields captured in prisma/schema.prisma's SensorReading model."""

    air_temperature: float = Field(..., description="Kelvin")
    process_temperature: float = Field(..., description="Kelvin")
    rotational_speed: float = Field(..., description="rpm")
    torque: float = Field(..., description="Nm")
    tool_wear: float = Field(..., description="minutes")
    machine_type: Literal["L", "M", "H"] = Field(
        ..., description="AI4I product quality variant (Low/Medium/High)"
    )


class PredictRequest(BaseModel):
    machine_id: str
    reading: SensorReadingIn


class PredictResponse(BaseModel):
    machine_id: str
    failure_probability: float
    threshold: float
    model_version: Optional[str] = None


class RulHistoryPoint(BaseModel):
    """One prior reading's (tool_wear, health_indicator) pair — both are needed to
    recompute the composite degradation index at that point in time."""

    tool_wear: float
    health_indicator: float = Field(..., ge=0.0, le=1.0, description="From /predict's failure_probability")


class RulRequest(BaseModel):
    machine_id: str
    tool_wear: float
    health_indicator: float = Field(..., ge=0.0, le=1.0, description="From /predict's failure_probability")
    # Prior readings for this machine, oldest first, so the caller doesn't need to
    # keep Wiener state itself — each call is stateless and recomputes from history.
    history: list[RulHistoryPoint] = Field(default_factory=list)


class RulResponse(BaseModel):
    machine_id: str
    current_degradation_index: float
    mu_posterior: float
    tau_posterior: float
    n_updates: int
    regime: Literal["prior-dominated", "data-informed", "machine-specific"]
    rul_p10_wear_minutes: float
    rul_p50_wear_minutes: float
    rul_p90_wear_minutes: float


class SynthesizeRequest(BaseModel):
    n_samples: int = Field(default=100, ge=1, le=10000)


class SynthesizeResponse(BaseModel):
    n_samples: int
    samples: list[list[float]]
    feature_columns: list[str]
    note: str = (
        "Generated by the Weibull-VAE for augmentation/visualization only — the notebook's "
        "own TSTR evaluation found a 0.125 F1 gap vs the real-data classifier, so these "
        "samples are not validated as a standalone training substitute."
    )


class HealthResponse(BaseModel):
    status: Literal["ok"]
    xgb_classifier_loaded: bool
    weibull_vae_loaded: bool
    scaler_loaded: bool