# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AfriMaint API",
    "version": "0.1.0",
    "description": "Contract for the Next.js API routes under app/api/. Frozen here so a future Django implementation (implementation plan Phase 6) can be validated against the same shapes before cutover."
  },
  "paths": {
...
    "/api/predict": {
      "post": {
        "summary": "Score a sensor reading via ml-service and persist the prediction",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "404": {
            "description": "Machine not found"
          },
          "502": {
            "description": "ML service unavailable"
          },
          "503": {
            "description": "ML service reachable but artifact not loaded"
          }
        }
      }
    },
    "/api/economics": {
      "get": {
        "summary": "List recent economic reports for the organization",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "summary": "Compute a run-to-failure vs predictive-maintenance scenario and persist it",
        "security": [
          {
            "sessionCookie": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EconomicScenario"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "404": {
            "description": "Machine not found"
          },
          "409": {
            "description": "No cost matrix configured"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token"
      }
    },
    "schemas": {
...
      "PredictRequest": {
        "$ref": "#/definitions/PredictRequest",
        "definitions": {
          "PredictRequest": {
            "type": "object",
            "properties": {
              "machineId": {
                "type": "string",
                "minLength": 1
              },
              "reading": {
                "type": "object",
                "properties": {
                  "airTemperature": {
                    "type": "number"
                  },
                  "processTemperature": {
                    "type": "number"
                  },
                  "rotationalSpeed": {
                    "type": "number"
                  },
                  "torque": {
                    "type": "number"
                  },
                  "toolWear": {
                    "type": "number"
                  },
                  "machineType": {
                    "type": "string",
                    "enum": [
                      "L",
                      "M",
                      "H"
                    ]
                  }
                },
                "required": [
                  "airTemperature",
                  "processTemperature",
                  "rotationalSpeed",
                  "torque",
                  "toolWear",
                  "machineType"
                ],
                "additionalProperties": false
              }
            },
            "required": [
              "machineId",
              "reading"
            ],
            "additionalProperties": false
          }
        }
      },
      "EconomicScenario": {
        "$ref": "#/definitions/EconomicScenario",
        "definitions": {
          "EconomicScenario": {
            "type": "object",
            "properties": {
              "machineId": {
                "type": "string",
                "minLength": 1
              },
              "idleWorkers": {
                "type": "number",
                "minimum": 0
              },
              "runToFailureDowntimeHours": {
                "type": "number",
                "minimum": 0
              },
              "predictiveDowntimeHours": {
                "type": "number",
                "minimum": 0
              },
              "preventiveMaintenanceCostKsh": {
                "type": "number",
                "minimum": 0
              }
            },
            "required": [
              "machineId",
              "idleWorkers",
              "runToFailureDowntimeHours",
              "predictiveDowntimeHours",
              "preventiveMaintenanceCostKsh"
            ],
            "additionalProperties": false
          }
        }
      }
    }
  }
}